backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0

# Document parsing
pypdf2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
chardet==5.2.0

# AI
google-generativeai==0.8.3
anthropic>=0.40.0

# Vector / Embeddings
pgvector==0.3.6
sentence-transformers==3.3.1

# HTTP
httpx==0.28.1
aiofiles==24.1.0
